Apply download quarantine on non-Mojang Java runtimes
Unlike the default sandbox quarantine, an executable file with a quarantine type that indicates that the file was downloaded is allowed to run. However, it is subject to Gatekeeper checks. Since Adoptium and Zulu are correctly signed and notarized, this isn't a problem and lets us offload the work to verifying their integrity to the OS. Some additional changes were made to how extraction is done to avoid invalidating the code signature.

Signed-off-by: Kenneth Chew <79120643+kthchew@users.noreply.github.com>